ai_backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
pyTelegramBotAPI
supabase
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
pypdf
python-dotenv
